+++ lib/td_dd/data_structures/bulk_updater.ex
@@ -74,13 +74,16 @@ defmodule TdDd.DataStructures.BulkUpdater do
       Task.Supervisor.async_nolink(
         TdDd.TaskSupervisor,
         fn ->
-          contents = BulkUpdate.from_csv(structures_content_upload)
-
-          {:ok, %{updates: updates, update_notes: update_notes}} =
-            BulkUpdate.do_csv_bulk_update(contents, user_id, auto_publish)
-
-          [updated_notes, not_updated_notes] = BulkUpdate.split_succeeded_errors(update_notes)
-          make_summary(updates, updated_notes, not_updated_notes)
+          with [_ | _] = contents <-
+                 BulkUpdate.from_csv(structures_content_upload),
+               {:ok, %{updates: updates, update_notes: update_notes}} <-
+                 BulkUpdate.do_csv_bulk_update(contents, user_id, auto_publish),
+               [updated_notes, not_updated_notes] <-
+                 BulkUpdate.split_succeeded_errors(update_notes) do
+            make_summary(updates, updated_notes, not_updated_notes)
+          else
+            {:error, _} = error -> error
+          end
         end
       )
 
@@ -92,12 +95,12 @@ defmodule TdDd.DataStructures.BulkUpdater do
       user_id: user_id,
       status: "STARTED",
       csv_hash: csv_hash,
-      task_reference: task.ref |> ref_to_string,
+      task_reference: ref_to_string(task.ref),
       filename: filename
     })
 
     %{
-      reply: {:just_started, csv_hash, task.ref |> ref_to_string},
+      reply: {:just_started, csv_hash, ref_to_string(task.ref)},
       state:
         put_in(
           state.tasks[task.ref],
@@ -160,6 +163,19 @@ defmodule TdDd.DataStructures.BulkUpdater do
   end
 
   # If the task succeeds...
+  @impl true
+  def handle_info({ref, {:error, error}} = msg, %{notify: notify} = state) do
+    # The task succeed so we can cancel the monitoring and discard the DOWN message
+    Process.demonitor(ref, [:flush])
+    {task_info, state} = pop_in(state.tasks[ref])
+
+    Process.cancel_timer(task_info.task_timer)
+
+    create_event(task_info, :DOWN, error)
+    maybe_notify(notify, msg)
+    {:noreply, state}
+  end
+
   @impl true
   def handle_info({ref, summary} = msg, %{notify: notify} = state) when is_reference(ref) do
     # The task succeed so we can cancel the monitoring and discard the DOWN message